Added more debugging, removed gl errors and slight optimizations
Added glObjectLabel to most objects, so in nsight i can know which object is what
removed most printfs that spam the console

removed the bind after a framebuffer blit
removed most of the code for calulating the skybox projection view code. the result should be the same
replaced the GetLastTransformUpdate function with a direct reference to it's variable

added a new GLDebug class that has the opengl error callback. the program should now capture and show all errors
using that, i fixed a few errors

moved the logging vertex count up past the skybox render so we dont include the skybox in it